Change relative ids to absolute as it's easier to handle

+++ contracts/types.sol
@@ -157,14 +157,14 @@ struct ActionInfo {
 
 // Equipment (leave at the bottom to allow for further ones)
 struct Armour {
-  uint8 helmet; // tokenId for the head (1 - 255)
-  uint8 amulet; // tokenId for the neck (256 - 511) (256 * i -> 256 * (i + 1))
-  uint8 chestplate;
-  uint8 gauntlets;
-  uint8 tassets;
-  uint8 boots;
-  uint8 reserved1;
-  uint8 reserved2;
+  uint16 helmet; // tokenId for the head (1 - 255)
+  uint16 amulet; // tokenId for the neck (256 - 511) (256 * i -> 256 * (i + 1))
+  uint16 chestplate;
+  uint16 gauntlets;
+  uint16 tassets;
+  uint16 boots;
+  uint16 reserved1;
+  uint16 reserved2;
 }
 
 // TODO: Can pack this better, remove the structs
